+++ tests/test_ngrok.py
@@ -407,3 +407,11 @@ class TestNgrok(NgrokTestCase):
         # THEN
         self.assertGreater(ngrok_tunnel.metrics.get("http").get("count"), 0)
         self.assertGreater(ngrok_tunnel.data["metrics"].get("http").get("count"), 0)
+
+    def test_version(self):
+        # WHEN
+        ngrok_version, pyngrok_version = ngrok.get_version()
+
+        # THEN
+        self.assertIsNotNone(ngrok_version)
+        self.assertEqual(ngrok.__version__, pyngrok_version)